refs #9785. Do not clean workspace that doesn't exit

+++ Code/Mantid/scripts/Reflectometry/isis_reflectometry/quick.py
@@ -228,7 +228,8 @@ def quick_explicit(run, i0_monitor_index, lambda_min, lambda_max,  background_mi
     
     if not debug:
         cleanup()
-        DeleteWorkspace('IvsLam')
+        if mtd.doesExist('IvsLam'):
+            DeleteWorkspace('IvsLam')
     return  mtd[RunNumber+'_IvsLam'], mtd[RunNumber+'_IvsQ'], theta
